+++ tests/test_pugsql.py
@@ -198,6 +198,13 @@ class PugsqlTest(TestCase):
             self.fixtures.find_by_username_or_id(
                 1, ('oscar', 'dottie'))
 
+    def test_mixed_positional_args_mistake(self):
+        with pytest.raises(
+                exceptions.InvalidArgumentError,
+                match='Pass keyword arguments to statements'):
+            self.fixtures.find_by_username_or_id(
+                1, usernames=('oscar', 'dottie'))
+
     def test_three_dashes(self):
         pytest.skip('fails - see issue #13')
         pugsql.module('tests/sql/extra-dashes')